% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_operator.R
\name{array_operator}
\alias{array_operator}
\title{Array operator to mimic different-dimension-array element-wise operations in
MATLAB. It receives as input two arrays of potentially different dimensions,
it resizes them to have same dimensions and finally performs the element-wise
operation.}
\usage{
array_operator(array1, array2, operation)
}
\arguments{
\item{array1}{The first array}

\item{array2}{The second array}

\item{operation}{The operation. It can take values: '+', '-', '*',
'/' and '^'}
}
\value{
An array with dimensions equal to the "largest" input array. It is
the result of applying the operator element-wise to both input arrays.
}
\description{
Array operator to mimic different-dimension-array element-wise operations in
MATLAB. It receives as input two arrays of potentially different dimensions,
it resizes them to have same dimensions and finally performs the element-wise
operation.
}
