% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_income_simple.R
\name{av_income_simple}
\alias{av_income_simple}
\title{Computes average income in each location, which is the weighted average of
the income of the people living in the location.}
\usage{
av_income_simple(lambda_ij_i, w_tr)
}
\arguments{
\item{lambda_ij_i}{NxN matrix - Probability of individuals in each
location of working in each location.}

\item{w_tr}{NxS - Wages in each location in each sector.}
}
\description{
Computes average income in each location, which is the weighted average of
the income of the people living in the location.
}
