% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuting_matrix.R
\name{commuting_matrix}
\alias{commuting_matrix}
\title{Function to transform travel times into iceberg commuting costs}
\usage{
commuting_matrix(t_ij, epsilon)
}
\arguments{
\item{t_ij}{NxN matrix - Travel time matrix across locations}

\item{epsilon}{Float - Parameter that transforms travel times to commuting costs}
}
\value{
A NxN matrix of commuting costs
}
\description{
Function to transform travel times into iceberg commuting costs
}
