% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivity.R
\name{productivity}
\alias{productivity}
\title{Computes productivity levels in each location}
\usage{
productivity(N, Q, w, L_j, K, t_ij, delta, lambda, beta)
}
\arguments{
\item{N}{Float - Number of locations.}

\item{Q}{Nx1 matrix - Floorspace prices in each location.}

\item{w}{Nx1 matrix - wages in each location.}

\item{L_j}{Nx1 matrix - Employment in each location.}

\item{K}{Nx1 matrix - Land in each location.}

\item{t_ij}{NxN matrix - Travel times matrix.}

\item{delta}{Float - decay parameter agglomeration.}

\item{lambda}{Float - agglomeration force.}

\item{beta}{Float - Output elasticity wrt labor}
}
\description{
Computes productivity levels in each location
}
