% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumDims2.R
\name{sumDims2}
\alias{sumDims2}
\title{Collapse array 2 along one of the dimensions by adding the elements along that
dimension.}
\usage{
sumDims2(array, dimension)
}
\arguments{
\item{array}{Array to collapse along one dimension.}

\item{dimension}{Dimension to collapse the array.}
}
\value{
An array that has been collapsed along the given dimension.
}
\description{
Collapse array 2 along one of the dimensions by adding the elements along that
dimension.
}
